Move some helper methods into out DebugClient
