+++ test/debug_client.ts
@@ -37,4 +37,42 @@ export class DartDebugClient extends DebugClient {
 		const thread = await this.getMainThread();
 		return this.stepInRequest({ threadId: thread.id });
 	}
+
+	public async getTopFrameVariables(scope: "Exception" | "Locals"): Promise<DebugProtocol.Variable[]> {
+		const thread = await this.getMainThread();
+		const stack = await this.stackTraceRequest({ threadId: thread.id });
+		const scopes = await this.scopesRequest({ frameId: stack.body.stackFrames[0].id });
+		const exceptionScope = scopes.body.scopes.find((s) => s.name === scope);
+		assert.ok(exceptionScope);
+		return this.getVariables(exceptionScope.variablesReference);
+	}
+
+	public async getVariables(variablesReference: number): Promise<DebugProtocol.Variable[]> {
+		const variables = await this.variablesRequest({ variablesReference });
+		return variables.body.variables;
+	}
+
+	public async evaluate(expression: string): Promise<{
+		result: string;
+		type?: string;
+		variablesReference: number;
+		namedVariables?: number;
+		indexedVariables?: number;
+	}> {
+		const thread = await this.getMainThread();
+		const stack = await this.stackTraceRequest({ threadId: thread.id });
+		const result = await this.evaluateRequest({ expression, frameId: stack.body.stackFrames[0].id });
+		return result.body;
+	}
+
+	public assertOutputContains(category: string, text: string) {
+		return new Promise((resolve, reject) => this.on("output", (event: DebugProtocol.OutputEvent) => {
+			if (event.body.category === category) {
+				if (event.body.output.indexOf(text) !== -1)
+					resolve();
+				else
+					reject(new Error(`Didn't find text "${text}" in ${category}`));
+			}
+		}));
+	}
 }
