+++ test/flutter_only/debug/flutter_run.test.ts
@@ -3,7 +3,7 @@ import * as vs from "vscode";
 import { DebugProtocol } from "vscode-debugprotocol";
 import { fsPath } from "../../../src/utils";
 import { DartDebugClient } from "../../debug_client";
-import { ensureVariable, getTopFrameVariables } from "../../debug_helpers";
+import { ensureVariable } from "../../debug_helpers";
 import { activate, defer, delay, ext, flutterHelloWorldBrokenFile, flutterHelloWorldFolder, flutterHelloWorldMainFile, getLaunchConfiguration, openFile, positionOf } from "../../helpers";
 
 describe("flutter run debugger", () => {
@@ -132,7 +132,7 @@ describe("flutter run debugger", () => {
 			dc.launch(config),
 		]);
 
-		const variables = await getTopFrameVariables(dc, "Exception");
+		const variables = await dc.getTopFrameVariables("Exception");
 		ensureVariable(variables, undefined, "message", `"(TODO WHEN UNSKIPPING)"`);
 	});
