+++ test/flutter_only/debug/flutter_test.test.ts
@@ -4,7 +4,6 @@ import * as vs from "vscode";
 import { DebugProtocol } from "vscode-debugprotocol";
 import { fsPath } from "../../../src/utils";
 import { DartDebugClient } from "../../debug_client";
-import { getTopFrameVariables } from "../../debug_helpers";
 import { activate, defer, ext, flutterHelloWorldFolder, flutterTestBrokenFile, flutterTestMainFile, flutterTestOtherFile, getLaunchConfiguration, openFile, positionOf } from "../../helpers";
 
 describe("flutter test debugger", () => {
@@ -158,7 +157,7 @@ describe("flutter test debugger", () => {
 			dc.launch(config),
 		]);
 
-		const variables = await getTopFrameVariables(dc, "Exception") as DebugProtocol.Variable[];
+		const variables = await dc.getTopFrameVariables("Exception") as DebugProtocol.Variable[];
 		assert.ok(variables);
 		const v = variables.find((v) => v.name === "message");
 		assert.ok(v);
